[NTS-UI/nui.calendar#36] :hammer: 팝업 UI 변경 및 컴포넌트 단일화
