+++ src/components/CalendarItem/CalendarItemPopup.module.scss
@@ -1,9 +1,8 @@
 .component {
 	position: absolute;
-	top: 0;
-	right: 0;
-	bottom: 0;
-	left: 0;
+	top: 50%;
+	left: 100%;
+	z-index: 2000;
 	&:before {
 		content: '';
 		position: fixed;
@@ -11,13 +10,11 @@
 		right: 0;
 		bottom: 0;
 		left: 0;
-		z-index: 2000;
 	}
-
 }
 .layer {
-	position: absolute;
-	z-index: 2000;
+	position: relative;
+	border-radius: 8px;
 	box-shadow: 0 0 2px 2px rgba(0, 0, 0, 0.1);
 	&:after {
 		content: '';
@@ -25,48 +22,19 @@
 		z-index: -1;
 		width: 10px;
 		height: 10px;
-		background-color: #fff;
+		left: -5px;
+		top: 50%;
+		background-color: #444;
 		box-shadow: 0 0 2px 2px rgba(0, 0, 0, 0.1);
 		transform: rotate(45deg);
 	}
 	&.hide {
 		display: none;
 	}
-	&.top {
-		bottom: 100%;
-		left: 0;
-		&:after {
-			bottom: -5px;
-			left: 7px;
-		}
-		&.left:after {
-			bottom: 5px;
-		}
-	}
-	&.bottom {
-		top: 100%;
-		left: 0;
-		&:after {
-			top: -5px;
-			left: 7px;
-		}
-		&.left:after {
-			top: 7px;
-		}
-	}
-	&.left {
-		right: 100%;
-		left: auto;
-		&:after {
-			left: auto;
-			right: -5px;
-		}
-	}
 	.inner {
 		width: 300px;
-		padding-top: 5px;
-		border-radius: 2px;
+		border-radius: 8px;
 		overflow: hidden;
-		background-color: #fff;
+		background-color: #444;
 	}
 }
