--- src/components/CalendarItem/CalendarItemPopupEditor.module.scss
@@ -1,54 +1,0 @@
-.component {
-    position: relative;
-    padding: 15px;
-    .button {
-        border: 0;
-        background-color: transparent;
-        -webkit-appearance: none;
-        appearance: none;
-        cursor: pointer;
-        &.lock {
-            margin-left: 4px;
-            padding: 6px;
-            border: solid 1px #999;
-            // TODO: 아이콘 변경
-            line-height: 10px;
-        }
-        &.submit {
-            display: inline-block;
-            vertical-align: top;
-            min-width: 96px;
-            font-size: 12px;
-            font-weight: 700;
-            line-height: 36px;
-            color: #fff;
-            background-color: #ff5252;
-            border-radius: 36px;
-        }
-        &.close {
-            position: absolute;
-            top: 0;
-            right: 0;
-            padding: 15px;
-        }
-    }
-    .row-info {
-        display: flex;
-        &:first-child {
-            padding-right: 33px;
-        }
-        + .row-info {
-            margin-top: 10px;
-        }
-    }
-    .delimiter {
-        margin: 0 2px;
-        line-height: 32px;
-    }
-    .area-button {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        margin-top: 10px;
-    }
-}
