+++ src/components/CalendarItem/Input.module.scss
@@ -1,9 +1,4 @@
 .item-input {
-    flex: 1;
-    position: relative;
-    height: 30px;
-    padding-left: 32px;
-    border: solid 1px #999;
     input {
         position: relative;
         border: 0;
@@ -12,9 +7,10 @@
         appearance: none;
         display: block;
         width: 100%;
-        height: 100%;
         font-size: 12px;
         line-height: 17px;
+        color: #fff;
+        outline: none;
         &::placeholder {
             color: #999;
         }
@@ -41,27 +37,24 @@
         display: inline-flex;
         align-items: center;
         vertical-align: top;
-        width: 40px;
-        height: 32px;
-        padding: 0 10px;
+        width: 20px;
+        height: 10px;
         border-radius: 16px;
         background-color: rgba(#bbb, 0.3);
         &:before {
             content: '';
             display: block;
-            width: 28px;
-            height: 28px;
-            margin-left: -8px;
+            width: 12px;
+            height: 12px;
             background-color: #fff;
             border-radius: 50%;
             box-shadow: 0 0 2px 2px rgba(#000, 0.05);
         }
     }
     :checked ~ label {
-        background-color: rgba(#09b65a, 0.3);
+        background-color: rgba(#09b65a, 0.7);
         &:before {
             margin-left: auto;
-            margin-right: -8px;
         }
     }
     input[type='checkbox'] {
@@ -74,18 +67,19 @@
     }
 }
 .item-select {
-    flex: 1;
     position: relative;
     select {
         position: relative;
         z-index: 1;
+        -webkit-appearance: none;
         appearance: none;
         width: 100%;
-        height: 30px;
-        box-sizing: border-box;
-        padding: 0 30px;
-        border: solid 1px #999;
+        border: 0;
         background-color: transparent;
+        font-size: 14px;
+        line-height: 18px;
+        color: #fff;
+        outline: none;
     }
     &:after {
         content: '';
@@ -95,22 +89,6 @@
         transform: translateY(-50%);
         border-style: solid;
         border-width: 4px;
-        border-color: #222 transparent transparent;
-    }
-}
-.icon {
-    position: absolute;
-    top: 50%;
-    left: 10px;
-    // TODO: 아이콘 변경
-    width: 10px;
-    height: 10px;
-    overflow: hidden;
-    transform: translateY(-50%);
-    &.category {
-        border-radius: 50%;
-    }
-    svg {
-        vertical-align: top;
+        border-color: #fff transparent transparent;
     }
 }
